# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * FastX Bus Ticket Booking API
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export type UserRegistrationGender = typeof UserRegistrationGender[keyof typeof UserRegistrationGender];


export const UserRegistrationGender = {
  male: 'male',
  female: 'female',
  other: 'other',
} as const;

export interface UserRegistration {
  email: string;
  /** @minLength 6 */
  password: string;
  name: string;
  gender: UserRegistrationGender;
  phone: string;
  /** @nullable */
  address?: string | null;
}

export interface OperatorRegistration {
  email: string;
  /** @minLength 6 */
  password: string;
  name: string;
  phone: string;
  companyName: string;
  /** @nullable */
  address?: string | null;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserProfileRole = typeof UserProfileRole[keyof typeof UserProfileRole];


export const UserProfileRole = {
  passenger: 'passenger',
  operator: 'operator',
  admin: 'admin',
} as const;

export interface UserProfile {
  id: number;
  email: string;
  name: string;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  role: UserProfileRole;
  /** @nullable */
  companyName?: string | null;
  createdAt?: string;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export interface UserProfileUpdate {
  name?: string;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  companyName?: string | null;
}

export type BusBusType = typeof BusBusType[keyof typeof BusBusType];


export const BusBusType = {
  seater_ac: 'seater_ac',
  seater_non_ac: 'seater_non_ac',
  sleeper_ac: 'sleeper_ac',
  sleeper_non_ac: 'sleeper_non_ac',
} as const;

export interface Amenities {
  waterBottle?: boolean;
  chargingPoint?: boolean;
  tv?: boolean;
  blanket?: boolean;
}

export interface Bus {
  id: number;
  name: string;
  busNumber: string;
  busType: BusBusType;
  totalSeats: number;
  operatorId: number;
  /** @nullable */
  operatorName?: string | null;
  amenities?: Amenities;
}

export type BusInputBusType = typeof BusInputBusType[keyof typeof BusInputBusType];


export const BusInputBusType = {
  seater_ac: 'seater_ac',
  seater_non_ac: 'seater_non_ac',
  sleeper_ac: 'sleeper_ac',
  sleeper_non_ac: 'sleeper_non_ac',
} as const;

export interface BusInput {
  name: string;
  busNumber: string;
  busType: BusInputBusType;
  /**
     * @minimum 10
     * @maximum 60
     */
  totalSeats: number;
  amenities?: Amenities;
}

export type RouteDetailStatus = typeof RouteDetailStatus[keyof typeof RouteDetailStatus];


export const RouteDetailStatus = {
  active: 'active',
  cancelled: 'cancelled',
  completed: 'completed',
} as const;

export interface RouteDetail {
  id: number;
  busId?: number;
  /** @nullable */
  busName?: string | null;
  /** @nullable */
  busNumber?: string | null;
  /** @nullable */
  busType?: string | null;
  origin: string;
  destination: string;
  departureTime: string;
  arrivalTime: string;
  fare: number;
  availableSeats: number;
  totalSeats: number;
  status: RouteDetailStatus;
  amenities?: Amenities;
  /** @nullable */
  operatorName?: string | null;
}

export interface RouteInput {
  busId: number;
  origin: string;
  destination: string;
  departureTime: string;
  arrivalTime: string;
  /** @minimum 0 */
  fare: number;
}

export interface Seat {
  id: number;
  seatNumber: string;
  isBooked: boolean;
  /** @nullable */
  seatRow?: number | null;
  /** @nullable */
  seatCol?: string | null;
}

export type BookingStatus = typeof BookingStatus[keyof typeof BookingStatus];


export const BookingStatus = {
  confirmed: 'confirmed',
  cancelled: 'cancelled',
  refunded: 'refunded',
} as const;

export interface Booking {
  id: number;
  userId: number;
  /** @nullable */
  userName?: string | null;
  routeId: number;
  /** @nullable */
  origin?: string | null;
  /** @nullable */
  destination?: string | null;
  /** @nullable */
  departureTime?: string | null;
  seatNumbers: string[];
  totalAmount: number;
  status: BookingStatus;
  createdAt: string;
  /** @nullable */
  busName?: string | null;
  /** @nullable */
  busNumber?: string | null;
}

export interface BookingInput {
  routeId: number;
  /** @minItems 1 */
  seatIds: number[];
}

export interface DashboardSummary {
  totalBookings: number;
  upcomingJourneys: number;
  cancelledBookings: number;
  totalSpent: number;
  recentBookings?: Booking[];
}

export interface RouteStats {
  routeId: number;
  origin: string;
  destination: string;
  bookingCount: number;
  revenue?: number;
}

export interface OperatorDashboard {
  totalBuses: number;
  totalRoutes: number;
  totalBookings: number;
  totalRevenue: number;
  pendingRefunds?: number;
  recentBookings?: Booking[];
  popularRoutes?: RouteStats[];
}

export interface AdminDashboard {
  totalUsers: number;
  totalOperators: number;
  totalRoutes: number;
  totalBookings: number;
  totalRevenue: number;
  recentBookings?: Booking[];
}

export type SearchRoutesParams = {
origin: string;
destination: string;
date: string;
};