# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: FastX Bus Ticket Booking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List bookings for current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Booking history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Book seats on a route
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Booking confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          description: Seats unavailable or validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all passengers (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProfile"
  /admin/users/{id}:
    delete:
      operationId: adminDeleteUser
      tags: [admin]
      summary: Delete a user account (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /admin/operators:
    get:
      operationId: adminListOperators
      tags: [admin]
      summary: List all bus operators (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All operators
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProfile"
  /admin/operators/{id}:
    delete:
      operationId: adminDeleteOperator
      tags: [admin]
      summary: Delete an operator account (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /admin/bookings:
    get:
      operationId: adminListBookings
      tags: [admin]
      summary: List all bookings (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
  /admin/dashboard:
    get:
      operationId: getAdminDashboard
      tags: [dashboard]
      summary: Admin dashboard summary
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Admin dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDashboard"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error

    # ── Auth ──────────────────────────────────────────────────────────────
...
    UserProfile:
      type: object
      required: [id, email, name, role]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        gender:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        role:
          type: string
          enum: [passenger, operator, admin]
        companyName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    Booking:
      type: object
      required: [id, userId, routeId, seatNumbers, totalAmount, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: ["string", "null"]
        routeId:
          type: integer
        origin:
          type: ["string", "null"]
        destination:
          type: ["string", "null"]
        departureTime:
          type: ["string", "null"]
          format: date-time
        seatNumbers:
          type: array
          items:
            type: string
        totalAmount:
          type: number
        status:
          type: string
          enum: [confirmed, cancelled, refunded]
        createdAt:
          type: string
          format: date-time
        busName:
          type: ["string", "null"]
        busNumber:
          type: ["string", "null"]
    BookingInput:
      type: object
      required: [routeId, seatIds]
      properties:
        routeId:
          type: integer
        seatIds:
          type: array
          items:
            type: integer
          minItems: 1

    # ── Dashboard ─────────────────────────────────────────────────────────
...
    AdminDashboard:
      type: object
      required: [totalUsers, totalOperators, totalRoutes, totalBookings, totalRevenue]
      properties:
        totalUsers:
          type: integer
        totalOperators:
          type: integer
        totalRoutes:
          type: integer
        totalBookings:
          type: integer
        totalRevenue:
          type: number
        recentBookings:
          type: array
          items:
            $ref: "#/components/schemas/Booking"
...