# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * FastX Bus Ticket Booking API
 * OpenAPI spec version: 1.0.0
 */
import type { BookingStatus } from './bookingStatus';

export interface Booking {
  id: number;
  userId: number;
  /** @nullable */
  userName?: string | null;
  routeId: number;
  /** @nullable */
  origin?: string | null;
  /** @nullable */
  destination?: string | null;
  /** @nullable */
  departureTime?: Date | null;
  seatNumbers: string[];
  totalAmount: number;
  status: BookingStatus;
  createdAt: Date;
  /** @nullable */
  busName?: string | null;
  /** @nullable */
  busNumber?: string | null;
}