# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * FastX Bus Ticket Booking API
 * OpenAPI spec version: 1.0.0
 */

export interface BookingInput {
  routeId: number;
  /** @minItems 1 */
  seatIds: number[];
}